adding local audio example
