+++ examples/Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
     "basic_room",
-    "local_audio_capture",
+    "local_audio",
     "mobile",
     "save_to_disk",
     "wgpu_room",
