+++ examples/local_audio/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "local_audio_capture"
+name = "local_audio"
 version = "0.1.0"
 edition = "2021"
